+++ src/Changes/MoreGates.cs
@@ -9,6 +9,7 @@ public class MoreGates : ManualChangesBase
   protected override void ApplyChangesInternal()
   {
     var hammer = ItemManager["Hammer"];
+    var pieceManager = this.GetPieceManager(hammer);
 
     // Relocate custom MoreGates pieces to appropriate categories
     Dictionary<string, Piece.PieceCategory> categoryChanges = new()
@@ -59,6 +60,34 @@ public class MoreGates : ManualChangesBase
     {
       var piece = prefab.GetComponent<Piece>();
       piece.m_category = categoryChanges[piece.name];
+    }
+
+    Plugin.Logger.LogDebug(hammer);
+    foreach (var piece in hammer.Pieces) Plugin.Logger.LogDebug(piece.name);
+
+    // Add iron to gate with metal
+    pieceManager["Hayze_gate_01"]?.Resources.Add("Iron", 1);
+
+    // Add crystal to window recipes
+    List<string> windows = [
+      "h_window_01",
+      "h_window_02",
+      "h_window_03",
+      "h_window_04",
+      "h_window_05",
+      "h_window_06",
+      "h_window_07",
+      "h_window_08",
+      "h_window_09",
+      "h_window_10",
+      "h_window_11",
+      "h_window_12",
+      "h_window_13",
+      "h_window_14",
+    ];
+    foreach (var prefabName in windows)
+    {
+      pieceManager[prefabName]?.Resources.Add("Crystal", 2);
     }
   }
 }